requirements.txt
# Core Dependencies
requests>=2.31.0
Pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# AI/ML - Modern stack
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0

# Discord
discord.py>=2.3.0

# Google Drive (updated for Python 3.12)
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.100.0

# Utilities
python-dotenv>=1.0.0